Moved logictree.to_toml inside valid.py
